refactor(watch): split the disk and redis watch modules into packages
Each was one file holding constants, helpers and a class, and disk also
held both halves of the feature: the delta-hook walk and the event hook
shared a file with nothing in common but the word watch.

  core/disk/watch/   constants  walk  hook
  core/redis/watch/  constants        hook

Same module names in both languages. The node barrels export only the
build hooks the resources import; tests reach the classes directly.
