+++ python/mirage/core/disk/watch/__init__.py
@@ -1,0 +1,18 @@
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+from mirage.core.disk.watch.hook import DiskEventHook, build_event_hook
+from mirage.core.disk.watch.walk import DiskWalk, build_delta_hook
+
+__all__ = ["DiskEventHook", "DiskWalk", "build_delta_hook", "build_event_hook"]
