+++ typescript/packages/node/src/core/disk/watch/constants.ts
@@ -1,0 +1,21 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import { FileChangeKind } from '@struktoai/mirage-core/types'
+
+export const DISK_KINDS: Record<string, FileChangeKind> = {
+  created: FileChangeKind.CREATE,
+  modified: FileChangeKind.UPDATE,
+  deleted: FileChangeKind.DELETE,
+}
