+++ typescript/packages/node/src/resource/redis/redis.ts
@@ -27,7 +27,7 @@ import { REDIS_COMMANDS } from '../../commands/builtin/redis/index.ts'
 import type { RedisClientType } from 'redis'
 import { RedisAccessor } from '../../accessor/redis.ts'
 import { type EventHook } from '@struktoai/mirage-core/watch/index'
-import { buildEventHook } from '../../core/redis/watch.ts'
+import { buildEventHook } from '../../core/redis/watch/index.ts'
 import { appendBytes } from '../../core/redis/append.ts'
 import { SCOPE_ERROR } from '../../core/redis/constants.ts'
 import { copy as copyCore } from '../../core/redis/copy.ts'
